Update: Use of DTOs

+++ LogicaAplicacion/DataTransferObjects/Mappers/MapperUsuario.cs
@@ -31,7 +31,7 @@ namespace LogicaAplicacion.DataTransferObjects.Mappers
             {
                 Email = user.Email.DireccionEmail,
                 Password = user.Contrasenia.Password,
-                Rol = user.Rol,
+                Rol = (int)user.Rol,
                 Nombre = user.NombreCompleto.Nombre,
                 Apellido = user.NombreCompleto.Apellido,
                 Id = user.Id
